requirements.txt
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.6
flake8==3.9.2
idna==3.2
influxdb==5.3.1
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.2
paho-mqtt==1.5.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
simpleeval==0.9.10
six==1.16.0
toml==0.10.2
urllib3==1.26.7
